Add support to detect invalid serializations.
This adds unit tests to catch a case of an invalid serialization.
It additionally outputs a better error message if someone uses
this with, say, a default "text/html" response and doesn't understand
why they then get a 500 error.

In addition, I've removed the Gemfile.lock file and added that to
.gitignore. The rubymine .idea directory has also been added to
.gitignore.