Retrieve ldp resource via subject with marmotta
This is some early concept code to retrieve a resource from a
configured backend (marmotta only at the moment). From what I read of
the used marmotta gem in this, it does caching when it pulls data for
the first time for a requested uri.

I've also added some initial configuration settings and some light unit
tests. These need to be expanded and made more error proof yet.

I've mocked out the http requests using VCR that is quite nifty. Can
set it to expire its cached files to require someone to rerun tests
after x times has passed? Or just good enough as it is?

Feel free to tear this code apart as I'm not entirely sure what I was
doing in all cases. :p