Rework, expand and document `BackendBase`
This substantially refactors the base backend interface, adds a
number of utility methods, and adds YARD annotations
throughout. Notably:

  - `#add` is implemented for cases when a resource should be added, but
     not retrieved.
  - `#delete_all` is deprecated in favor of `#delete_all!`
  - `#empty?` is provided.
  - `#has_resource?` is added to check if a resource is in the backend.

The `#add` and `#has_resource` methods are generally already implemented
as part of `#retrieve`, so this may be a first step toward a generic
implementation of that key method.

Shared examples are added, and tested against a completed `Repository`
implementation. Further work is needed to bring `Marmotta` and
`Blazegraph` backends up to date.

Some todo items (questions) are added for `#get_resource_uri`.