Setup Guard
Adds a generic Guardfile.

`spec_helper.rb` includes a stub on `Setting` in a before(:each), which
prevents various spec files from running without the rest of the
suite. This pattern also has the issue that it stubs
`Setting#cache_backend` without warning to developers, giving unexpected
results. We should find a way to remove it. See #38.

In the meanwhile, we introduce a guard to prevent errors, and allow
modules that don't depend on `Setting` to be tested independently.